+++ Source/Samples/0_TestZone/Src/AppRenderer.cpp
@@ -40,7 +40,7 @@ void AppRenderer::Initialize() {
   HEAP_ALLOCATOR(Temporary, Memory::MonotonicAllocator, 16384);
   m_window = RenderSystem::CreateApplicationWindow("TestZone", 1280, 720);
 
-  m_window->SetUpdateCallback([this](double timeSinceLastFrame)
+  m_window->SetUpdateCallback([this](float timeSinceLastFrame)
   {
     UNUSED(timeSinceLastFrame);
     WindowUpdate();
