+++ Source/Samples/1_ProceduralPlanet/Src/AppRenderer.cpp
@@ -56,7 +56,7 @@ void AppRenderer::Initialize() {
   HEAP_ALLOCATOR(Temporary, Memory::MonotonicAllocator, 8192);
   m_window = RenderSystem::CreateApplicationWindow("ProceduralPlanet", 1280, 720);
 
-  m_window->SetUpdateCallback([this](double timeSinceLastFrame)
+  m_window->SetUpdateCallback([this](float timeSinceLastFrame)
   {
     UNUSED(timeSinceLastFrame);
     WindowUpdate();
